Move System.Native/pal_tcpstate.cpp to C

+++ src/Native/Unix/System.Native/pal_networkstatistics.cpp
@@ -300,7 +300,7 @@ extern "C" int32_t SystemNative_GetActiveTcpConnectionInfos(NativeTcpConnectionI
         in_pcb = head_xtcpb->xt_inp;
 
         NativeTcpConnectionInformation* ntci = &infos[connectionIndex];
-        ntci->State = SystemNative_MapTcpState(tcp_pcb.t_state);
+        ntci->State = static_cast<int32_t>(SystemNative_MapTcpState(tcp_pcb.t_state));
 
         uint8_t vflag = in_pcb.inp_vflag; // INP_IPV4 or INP_IPV6
         bool isIpv4 = (vflag & INP_IPV4) == INP_IPV4;
